Split EditPanel and CompactView into separate comp
