+++ src/components/InRequest/InRequest.tsx
@@ -3,7 +3,6 @@ import {
   DatePicker,
   IComboBox,
   IComboBoxOption,
-  Panel,
 } from "@fluentui/react";
 import {
   ChangeEvent,
@@ -14,7 +13,6 @@ import {
   useState,
 } from "react";
 import { PeoplePicker, SPPersona } from "../PeoplePicker/PeoplePicker";
-import { useBoolean } from "@fluentui/react-hooks";
 import { OFFICES } from "../../constants/Offices";
 import { GS_GRADES, NH_GRADES, MIL_GRADES } from "../../constants/GradeRanks";
 import { emptype, empTypeOpts, EMPTYPES } from "../../constants/EmpTypes";
@@ -28,12 +26,10 @@ import {
   Radio,
   RadioGroup,
   RadioGroupOnChangeData,
-  Text,
-  webLightTheme,
-  FluentProvider,
 } from "@fluentui/react-components";
 import { UserContext } from "../../providers/UserProvider";
 import { IInForm, INewInForm, RequestApiConfig } from "../../api/RequestApi";
+import { InRequestViewCompact } from "./InRequestViewCompact";
 
 /**
  * Enum for holding the possible views of the In Request form view
@@ -52,19 +48,11 @@ export enum INFORMVIEWS {
 /* FluentUI Styling */
 const useStyles = makeStyles({
   formContainer: { display: "grid", paddingLeft: "1em", paddingRight: "1em" },
-  compactContainer: {
-    display: "grid",
-    paddingLeft: "1em",
-    paddingRight: "1em",
-    gridTemplateColumns: "repeat(auto-fit, minmax(150px,1fr))",
-    gridAutoRows: "minmax(50px, auto)",
-  },
   floatRight: {
     float: "right",
   },
 });
 
-/* TODO - Break into subcomponents for different views */
 export const InRequest: FunctionComponent<any> = (props) => {
   const classes = useStyles();
   const requestApi = RequestApiConfig.getApi();
@@ -83,43 +71,13 @@ export const InRequest: FunctionComponent<any> = (props) => {
     supGovLead: undefined,
   };
 
-  const [formData, setFormData] = useState<INewInForm | IInForm>(defaultInForm);
-  const [saveCancelEvent, setSaveCancelEvent] = useState<string>("");
-
-  /* Boolean state for determining whether or not the Edit Panel is shown */
-  const [isEditPanelOpen, { setTrue: showEditPanel, setFalse: hideEditPanel }] =
-    useBoolean(false);
-
-  /* Function to handle the save or cancel */
-  const onEditSaveCancel = (formEdits: IInForm | undefined): void => {
-    // If they chose to save the Edits, then update the formData
-    if (formEdits) {
-      setFormData({ ...formEdits });
-    }
-    hideEditPanel();
-    setSaveCancelEvent("");
-  };
+  const [formData, setFormData] = useState<INewInForm | IInForm>(
+    props.formData ? props.formData : defaultInForm
+  );
 
   const [gradeRankOptions, setGradeRankOptions] = useState<IComboBoxOption[]>(
     []
   );
-
-  const displayEmpType = (): string => {
-    let displayValue = "";
-    switch (formData.empType) {
-      case EMPTYPES.CIV:
-        displayValue =
-          "Civilian - " + (formData.isNewCiv === "yes" ? "New" : "Existing");
-        break;
-      case EMPTYPES.MIL:
-        displayValue = "Military";
-        break;
-      case EMPTYPES.CTR:
-        displayValue = "Contractor";
-        break;
-    }
-    return displayValue;
-  };
 
   const updateGradeRankOpt = (empType: emptype) => {
     if (empType) {
@@ -143,10 +101,6 @@ export const InRequest: FunctionComponent<any> = (props) => {
           return { ...f, prevOrg: "" };
         });
       }
-
-      setFormData((f: INewInForm) => {
-        return { ...f, empType: empType as emptype };
-      });
     }
   };
 
@@ -155,7 +109,9 @@ export const InRequest: FunctionComponent<any> = (props) => {
     data: RadioGroupOnChangeData
   ) => {
     let empType: emptype = data.value as emptype;
-    updateGradeRankOpt(empType);
+    setFormData((f: INewInForm) => {
+      return { ...f, empType: empType as emptype };
+    });
   };
 
   const onWorkLocationChange = (
@@ -244,33 +200,6 @@ export const InRequest: FunctionComponent<any> = (props) => {
     });
   };
 
-  const onRenderFooterContent = () => (
-    <FluentProvider theme={webLightTheme}>
-      <div>
-        <Button
-          appearance="primary"
-          onClick={() => {
-            setSaveCancelEvent("save");
-          }}
-        >
-          Save
-        </Button>
-        <Button
-          appearance="secondary"
-          onClick={() => {
-            setSaveCancelEvent("cancel");
-          }}
-        >
-          Cancel
-        </Button>
-      </div>
-    </FluentProvider>
-  );
-
-  function editRecord() {
-    showEditPanel();
-  }
-
   useEffect(() => {
     // Only set the supGovLead if this is a New request
     if (props.view === INFORMVIEWS.NEW) {
@@ -307,114 +236,20 @@ export const InRequest: FunctionComponent<any> = (props) => {
       const res = await requestApi.getItemById(props.ReqId);
       if (res) {
         setFormData(res);
-
-        //Update the available grade options dropdown based on the incoming record
-        updateGradeRankOpt(res?.empType);
       }
     };
 
     loadRequest();
   }, [props.ReqId, requestApi]);
 
+  useEffect(() => {
+    updateGradeRankOpt(formData?.empType as emptype);
+  }, [formData.empType]);
+
   /* Temporarily show a Loading screen if we don't have the current user info yet. */
   if (userContext.loadingUser) {
     return <>Loading...</>;
   }
-
-  const editPanel = (
-    <Panel
-      isOpen={isEditPanelOpen}
-      isBlocking={true}
-      onDismiss={hideEditPanel}
-      headerText="Edit Request"
-      isFooterAtBottom={true}
-      onRenderFooterContent={onRenderFooterContent}
-    >
-      <FluentProvider theme={webLightTheme}>
-        {/* TODO - Determine if this
-        compoent should retrieve the form data again from the source, or pass it
-        what is currently on the currently loaded form. For now it is reading in
-        the form data again from the source */}
-        {/* Recursively embed the component inside */}
-        <InRequest
-          ReqId={props.ReqId}
-          view={INFORMVIEWS.EDIT}
-          onEditSaveCancel={onEditSaveCancel}
-          saveCancelEvent={saveCancelEvent}
-        />
-      </FluentProvider>
-    </Panel>
-  );
-
-  /* This view renders a compact view of the items as simple Text elements (not fields) */
-  const compactView = (
-    <>
-      <div id="inForm" className={classes.compactContainer}>
-        <div>
-          <Label weight="semibold" htmlFor="empNameId">
-            Employee Name:
-          </Label>
-          <br />
-          <Text id="empNameId">{formData.empName}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="empTypeId">
-            Employee Type
-          </Label>
-          <br />
-          <Text id="empTypeId">{displayEmpType}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="gradeRankId">
-            Grade/Rank
-          </Label>
-          <br />
-          <Text id="gradeRankId">{formData.gradeRank}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="workLocationId">
-            Local or Remote?
-          </Label>
-          <br />
-          <Text id="workLocationId">{formData.workLocation}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="arrivalDateId">
-            Select estimated on-boarding date
-          </Label>
-          <br />
-          <Text id="arrivalDateId">{formData.eta?.toLocaleDateString()}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="officeId">
-            Office
-          </Label>
-          <br />
-          <Text id="officeId">{formData.office}</Text>
-        </div>
-        <div>
-          <Label weight="semibold" htmlFor="supGovLeadId">
-            Supervisor/Government Lead
-          </Label>
-          <br />
-          <Text id="supGovLeadId">{formData.supGovLead?.text}</Text>
-        </div>
-        {formData.empType === EMPTYPES.CIV && formData.isNewCiv === "no" && (
-          <div>
-            <Label weight="semibold" htmlFor="prevOrgId">
-              Previous Organization
-            </Label>
-            <br />
-            <Text>{formData.prevOrg}</Text>
-          </div>
-        )}
-      </div>
-      <Button appearance="primary" className="floatRight" onClick={editRecord}>
-        Edit
-      </Button>
-      {editPanel}
-    </>
-  );
 
   /* This view serves up the form fields as editable.  It is utilied by both NEW and EDIT forms */
   const formView = (
@@ -537,7 +372,7 @@ export const InRequest: FunctionComponent<any> = (props) => {
   const selectedView = (() => {
     switch (props.view) {
       case INFORMVIEWS.COMPACT:
-        return compactView;
+        return <InRequestViewCompact formData={formData as IInForm} />;
       case INFORMVIEWS.NEW:
       case INFORMVIEWS.EDIT:
       default:
